Moves trainer into separate file, reorganizes learn
